Fixes bnc#880927
The `cmd` API treats anything truish returned from a command as an
instruction to quit the program in `cmd.Cmd.cmdloop`.

This API has been applied inconsistently and some commands return truish
stuff when they should not but the program doesn't actually exit
because:

1. the cmdloop is ran inside `while True` cycle so it just starts the
   cmdloop again.

2. quiting is performed by calling sys.exit directly instead of exiting
   via the cmd API.

3. `cmd.Cmd` preloop/postloop features were never used for anything
   (which could be confused by unexpectedly restarting the cmdloop).

Commit 3390bcf introduced CommandPrompt.postloop that raises QuitLoop to
propagate the "quit instruction" to the outer loop. However that led to
the discovery of the commands with bad returns.

Therefore this propagation needs to be reverted until the relevant
commands are refactored to not return other values.