Add config option mtui.use_keyring
to controll wheter try to connect to keyring so this is not determined
just by wheter the keyring is present or not which may cause confusion
for users who get keyring installad by accident. For example by pulling
it by as Recommended package.

Or users might want to temporary disable keyring for whatever reason
(debugging, testing, temporary breakage in keyring, etc)