Add special case for "sle" product
Currently "sle" attribute actually doesn't exist and if used, is ignored.

This makes life a little difficult as asking for `sle 12` will choose
opensuse 12.x systems as well. So this commit adds special case so
instead of

    xml_p == required_p

it's compared as

    xml_p without last character == required_p

which will turn "sled" and "sles" from xml_p into "sle" and will match.