Add test for BSC#911686
- list_metadata missing tested architectures info

Since the relation TestReport:TestPlatforms is 1:N and I want to display
a testplatform per line for good readability, I expect one key
("Testplatform") to appear multiple times for simplicity of
implementation and the output format.

Therefore the TestReport._aligned_write() needs to take data as a
2-tuple list instead of dict.

The definition of Testplatforms in tests for md5 and rrid testreports is
inconsistent but the rrid one is much shorter therefore I opted to keep
it in one string.