Refactor searching
in `Refhosts.search` and `Refhosts.get_host_attributes`
using the `Refhsots._location_hosts` helper.

The old code checked for AssertionError and IndexError.

AssertionError seemed to indicate condition where no hosts were found in
the try-block search. Therefore was replaced with a simple
``if hosts == []``.

IndexError was probably to catch the condition where self.location was
invalid but that is no longer a concern since it is checked in both
`set_location` command and when reading the location in mtui.config.