Improve run command parsing
with regard to bsc#932002.

This is probably the best implementation given the current interface
design. The command just takes hostnames from the start until a hostname
is not matched and then the rest of arguments is treated as command.

Note this makes impossible to detect when a wrong hostname is given. It
will be regarded simply as part of the command.

This is legacy code hotfix and as such is tested only manually.

N valid hosts, "," in command::

    mtui> run fletcher.qam.suse.de, cunningham.qam.suse.de,echo "foo,bar"
    cunningham.qam.suse.de:~> echo "foo,bar" [0]
    foo,bar

    fletcher.qam.suse.de:~> echo "foo,bar" [0]
    foo,bar

    info: done

N valid hosts, valid host enclosed in "," in command::

    mtui> run fletcher.qam.suse.de, cunningham.qam.suse.de,echo "foo,merope.qam.suse.de,bar"
    cunningham.qam.suse.de:~> echo "foo,merope.qam.suse.de,bar" [0]
    foo,merope.qam.suse.de,bar

    fletcher.qam.suse.de:~> echo "foo,merope.qam.suse.de,bar" [0]
    foo,merope.qam.suse.de,bar

    info: done

"all" hosts, valid host enclosed in "," in command::

    mtui> run all ,echo "foo,merope.qam.suse.de,bar"
    cunningham.qam.suse.de:~> echo "foo,merope.qam.suse.de,bar" [0]
    foo,merope.qam.suse.de,bar

    merope.qam.suse.de:~> echo "foo,merope.qam.suse.de,bar" [0]
    foo,merope.qam.suse.de,bar

    fletcher.qam.suse.de:~> echo "foo,merope.qam.suse.de,bar" [0]
    foo,merope.qam.suse.de,bar

    info: done

Double comma separating hosts and command::

    mtui> run all ,,echo "foo,merope.qam.suse.de,bar"
    cunningham.qam.suse.de:~> ,echo "foo,merope.qam.suse.de,bar" [127]

    stderr:
    bash: ,echo: command not found

The last case is left this way as fixing it would at the same time break
commands with ",," substring.