check_new_depencencies.sh, compare_new_dependencies.sh anew
the old code had two problems:

* it disabled all repo's with TESTING in name, causing problems for
  updates without proper rep-clean support such as SUSE-Cloud.

  let's say you need to test something rep-clean.sh does not know or
  care about.  you'll have to add the repository with candidate packages
  by hand, and it needs to be named "TESTING-$ISSUEID".  then `update`
  in mtui will attempt to reset repositories into a known state for you
  with `rep-clean.sh -z -i $ISSUEID` and fail, leaving you without the
  repo you snuck in (`-z` works well, it's the -`i $ISSUEID` part that
  fails).

* it was inefficient and worked in an obscure and roundabout manner.

  it was *in theory* a more thorough approach (compare lists of all
  installed packages that *should* be from the TESTING repositories),
  but with shared refhosts all bets are off, and you never know if
  a result you've got isn't tainted by a concurrent work of another
  tester.

the new code does not touch any repositories.  dependencies for each
package under test are collected with `rpm -q --requires`, each line
tagged with the name of the dependent package.  since this observes
package metadata instead of actual effects, there's a possibility
a rogue dependency might sneak in, maybe through malevolent or
incompetent code in a %post hook or similar.  there's a concensus
at least among osukup@, rneuhauser@ and tchvatal@ that this should
not be a concern for this particular script.