fix downgrade command with multiple hosts, different states
* broken in d8bcb0becdd0991e1d87ed17948735cde7906d9d
* fixes bsc#939198

mtui avoids issuing commands to downgrade packages which are part
of the update but not installed on a particular target.  to do so,
mtui copies the HostsGroup and then prunes it with `del temp[hn]`.
alas, HostsGroup was missing __delitem__ and
CommandPrompt#do_downgrade was "kind enough" to eat the exception
and emit an information-poor error message.