HostsGroupException gone
HostsGroupException's purpose was to let all targets have a go
at massaging themselves even in presence of exceptions:
the issue was/is that HostsGroup has had no access to logging.
fixing that then involved a lot of uncertainty so we went for
the smaller source change instead, at the price of lost interface
simplicity.

the only use of HostsGroupException was in HostsGroup#unlock,
where it filtered out TargetLockedError exceptions and re-raised
a new HostsGroupException with the rest, losing the original
HostsGroupException's stacktrace in the process.

dropping this contraption, we're left with this state:

* Target#unlock logs TargetLockedError instances before they're let out
* HostsGroup#unlock catches and ignores TargetLockedError

this means that a bug in mtui won't have you waiting for all targets
in a doomed unlock command, and exceptions which are meant to escape
HostsGroup#unlock will finally have useful stacktraces.