works with both rep-clean and repose
`export` does not bother matching exact commands anymore,
it just replaces everything below "put here the output (...)"
with the actual commands and their inputs; this means `export`
will work more reliably.

the commands performed by mtui in the course of updating had to
be modified: only those command lines that start with "zypper "
are included; this is to keep the auxiliary commands (`prepare`,
checks) and their outputs from ending up in the log.