Query Package versions before printing any warning to user after a downgrade
The packages for each host, are now queried before printing any warning to the user
after a downgrade. After the querying the package's "after" and "before" properties
are updated. Without these changes when there was a package that wasn't installed
and the user would do a "update --prepare" and later a "downgrade" it would get a
warning because of the package that wasn't updated.