Hardcoded the timezone in the locks timestamp string
The lock timestamps are using the function time.time() to store the timestamp information.
This function returns the number of seconds since epoch (in UTC), making not
required to save the timezone information.

Since the users are requiring the information of the locks timezone to be displayed
the timezone information is hardcoded