Cleanup repeated information from the template
This commit solves two issues:
1- Repeated line "After:\nAfter:\n" by looping through all lines and
checking if the current line is the same as the previous. If that's
the case, then skips it.

2- Repeated scripts installation outcome information.
This case is more complex as there are 4 possible cases:
a- When the user does several downgrades and updates in the same session
b- When the user does a downgrade/update, exits and then in another session
   does again the another downgrade/update
c- When the user does a downgrade/update exits and then in another session
   doesn't do anything and just exports
d- When the user does a downgrade/update, exits and then in another session
   adds different hosts and does again the another downgrade/update

This is done in two parts. The first is when loading the template file from
the disk. If it contains the script outcome and the host of the new data is
present then deletes it from the loaded data of the template.

The second part, is because the xmldata parameter contains all the history
done to certain host in the current session and we only want the last one.
So for each host, we store the line position for the last script entry in
the template and then we check if it contains the output for that script and
we act accordingly.